backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-cohere==0.1.5
openai==1.30.1
cohere==5.5.4
tiktoken==0.7.0
faiss-cpu==1.8.0